针对设备新增任务处理

+++ HealthMonitor/Model/Entity/PassengerScreenEntity.cs
@@ -37,7 +37,7 @@ namespace HealthMonitor.Model.Entity
         /// <summary>
         /// 推送间隔
         /// </summary>
-        public decimal PUSHINTERVAL { get; set; }
+        public double PUSHINTERVAL { get; set; }
 
         /// <summary>
         /// 关联区域编号
